server: fix settings loss
When writing the settings write them with the sync flag
to ensure that the data is written to storage which may be
on a remote system. This should prevent issues where the
process exits after saving the settings but the data has not
been flushed from the buffers. To prevent too much slow down
from this cache the last wrote settings and only write when
they change.